refactor: split palettes.rs into per-palette modules
Mechanical split of 853-line palettes.rs into palettes/ directory.
Each theme palette gets its own file. No value changes.
